move context to `/snaps/`
