+++ ui/contexts/snaps/utils.ts
